plot_dmat {pbdDEMO}	R Documentation

Visualizing the DMAT Data Structure

Description

Plot a (small) global matrix as though it had been chopped up into pieces in the block-cyclic fashion.

Usage

plot_dmat(nrow, ncol, nprow, npcol, bldim, ..., labeling = "blacs",
  col = "rainbow")

Arguments

nrow, ncol
Number of global rows/columns of the matrix.

nprow, npcol
Number of processor rows/columns in the BLACS grid.

bldim
The blocking factor for the data distribution.

...
Additional arguments

labeling
Character argument; should be "blacs" or "mpi". This determines how the processor labeling should be, either in the 2-d BLACS way, or in the 1-d MPI way.

col
R plots color argument

Details

This function helps the user visualize 2-d block-cyclic distributed data.

[Package pbdDEMO version 0.3-0 Index]